+++ R/n2k_hurdle_imputed_class.R
@@ -1,30 +1,34 @@
-#' The `n2kImport` class
+#' The `n2kHurdleImputed` class
 #'
-#' It hold the imported datasets
-#' @slot Dataset A `data.frame` with the datasets.
-#' @name n2kImport-class
-#' @rdname n2kImport-class
-#' @exportClass n2kImport
+#' It holds the model2 of hurdle imputed data.
+#' @slot Presence A `rawImputed` object with the presence component.
+#' @slot Count A `rawImputed` object with the count component.
+#' @slot Hurdle A `aggregatedImputed` object with the resulting hurdle object.
+#' @name n2kHurdleImputed-class
+#' @rdname n2kHurdleImputed-class
+#' @exportClass n2kHurdleImputed
+#' @aliases n2kHurdleImputed-class
 #' @importFrom methods setClass
 #' @docType class
-#' @include n2k_model_class.R
 setClass(
-  "n2kImport",
+  "n2kHurdleImputed",
   representation = representation(
-    Dataset = "data.frame"
+    Presence = "maybeRawImputed", Count = "maybeRawImputed",
+    Hurdle = "maybeAggregatedImputed"
   ),
   contains = "n2kModel"
 )
 
 #' @importFrom methods setValidity
 #' @importFrom digest sha1
-#' @importFrom assertthat assert_that has_name
+#' @importFrom assertthat assert_that is.string
 setValidity(
-  "n2kImport",
+  "n2kHurdleImputed",
   function(object) {
-    assert_that(has_name(object@Dataset, "fingerprint"))
-    assert_that(has_name(object@Dataset, "filename"))
-    assert_that(has_name(object@Dataset, "import_date"))
+    assert_that(
+      nrow(object@AnalysisRelation) == 2,
+      msg = "`n2kHurdleImpute` requires exactly two parent analyses"
+    )
 
     file_fingerprint <- sha1(
       list(
@@ -38,8 +42,7 @@ setValidity(
         object@AnalysisMetadata$duration,
         object@AnalysisMetadata$last_analysed_year,
         format(object@AnalysisMetadata$analysis_date, tz = "UTC"),
-        object@AnalysisMetadata$seed,
-        object@AnalysisRelation$parent_analysis
+        object@AnalysisMetadata$seed, object@AnalysisRelation$parent_analysis
       ),
       environment = FALSE
     )
@@ -52,16 +55,15 @@ setValidity(
     status_fingerprint <- sha1(
       list(
         object@AnalysisMetadata$file_fingerprint,
-        object@AnalysisMetadata$status,
-        object@AnalysisMetadata$analysis_version, object@AnalysisVersion,
-        object@RPackage, object@AnalysisVersionRPackage,
-        object@Dataset
+        object@AnalysisMetadata$status, object@AnalysisVersion, object@RPackage,
+        object@AnalysisVersionRPackage, object@AnalysisRelation,
+        object@Presence, object@Count, object@Hurdle
       ),
       digits = 6L
     )
 
     assert_that(
-      object@AnalysisMetadata$status_fingerprint == status_fingerprint,
+      object@AnalysisMetadata$status_fingerprint != status_fingerprint,
       msg = "Corrupt status_fingerprint"
     )