+++ tests/testthat/helper_test_data.R
@@ -1,5 +1,5 @@
 #' A function to generate a simple dataset for unit testing
-#' @param datasource_id the string id of the datasource
+#' @param datasource_id the string id of the data source
 #' @param missing the required proportion of missing data
 test_data <- function(datafield_id = sha1(letters), missing = 0) {
   assertthat::assert_that(assertthat::is.string(datafield_id))