+++ tests/testthat/test_bbb_store_manifest_yaml.R
@@ -1,7 +1,7 @@
 context("store_manifest_yaml")
 test_that("store_manifest_yaml stores the manifest on an S3 bucket", {
   skip_if(Sys.getenv("AWS_SECRET_ACCESS_KEY") == "", message = "No AWS access")
-  bucket <- get_bucket("n2kmonitoring")
+  bucket <- get_bucket(Sys.getenv("N2KBUCKET"))
   object <- n2k_manifest(
     data.frame(
       fingerprint = "1", parent = NA_character_, stringsAsFactors = FALSE
@@ -9,7 +9,7 @@ test_that("store_manifest_yaml stores the manifest on an S3 bucket", {
   )
   project <- "unittest_store_manifest_yaml"
   docker <- "inbobmk/rn2k:latest"
-  dependencies <- c("inbo/n2khelper@v0.4.1", "inbo/n2kanalysis@docker") # nolint: nonportable_path_linter, line_length_linter.
+  dependencies <- c("inbo/n2khelper@v0.4.1", "inbo/n2kanalysis@docker")
   expect_is(
     stored <- store_manifest_yaml(
       x = object, base = bucket, project = project, docker = docker,
@@ -57,7 +57,7 @@ test_that("store_manifest_yaml stores the manifest on a file system", {
   )
   project <- "unittest_store_manifest_yaml"
   docker <- "inbobmk/rn2k:latest"
-  dependencies <- c("inbo/n2khelper@v0.4.1", "inbo/n2kanalysis@docker") # nolint: nonportable_path_linter, line_length_linter.
+  dependencies <- c("inbo/n2khelper@v0.4.1", "inbo/n2kanalysis@docker")
   expect_is(
     stored <- store_manifest_yaml(
       x = object, base = base, project = project, docker = docker,